took sidebar out
