+++ frontend/src/lib/to-proper-case.ts
@@ -1,0 +1,6 @@
+export default function toProperCase(s: string) {
+  return s.replace(
+    /\w\S*/g,
+    (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase()
+  )
+}
